# Use an approximation of the heightmap derivative to "erode" some layers of ground_structure