--- a/assets/cubyz/biomes/tall_mountain/peak.zig.zon
+++ b/assets/cubyz/biomes/tall_mountain/peak.zig.zon
@@ -7,7 +7,7 @@
 	.interpolation = .linear,
 	.mountains = 30,
 	.chance = 0,
-	
+
 	.erosion = 1.5,
 
 	.music = "cubyz:out_of_breath",