--- a/assets/cubyz/biomes/tall_mountain/peak.zig.zon
+++ b/assets/cubyz/biomes/tall_mountain/peak.zig.zon
@@ -8,7 +8,7 @@
 	.mountains = 30,
 	.chance = 0,
 
-	.erosion = 1.5,
+	.soilCreep = 1.5,
 
 	.music = "cubyz:out_of_breath",