--- a/src/server/terrain/biomes.zig
+++ b/src/server/terrain/biomes.zig
@@ -466,7 +466,7 @@ pub const BlockStructure = struct { // MARK: BlockStructure
 
 	pub fn addSubTerranian(self: BlockStructure, chunk: *ServerChunk, startingDepth: i32, minDepth: i32, slope: f32, erosion: f32, x: i32, y: i32, seed: *u64) i32 {
 		var depth = startingDepth;
-		var timer = @max(@as(i32, @intFromFloat(slope*erosion))-1, 0);
+		var timer = @max(@as(i32, @intFromFloat(slope*erosion)) - 1, 0);
 		for(self.structure) |blockStack| {
 			const total = blockStack.min + main.random.nextIntBounded(u32, seed, @as(u32, 1) + blockStack.max - blockStack.min);
 			for(0..total) |_| {