--- a/src/server/terrain/biomes.zig
+++ b/src/server/terrain/biomes.zig
@@ -289,7 +289,7 @@ pub const Biome = struct { // MARK: Biome
 	caves: f32,
 	caveRadiusFactor: f32,
 	crystals: u32,
-	erosion: f32, // How much of the surface structure should be eroded depending on the slope.
+	soilCreep: f32, // How much of the surface structure should be eroded depending on the slope.
 	stoneBlock: main.blocks.Block,
 	fogLower: f32,
 	fogHigher: f32,
@@ -335,7 +335,7 @@ pub const Biome = struct { // MARK: Biome
 			.caves = zon.get(f32, "caves", -0.375),
 			.caveRadiusFactor = @max(-2, @min(2, zon.get(f32, "caveRadiusFactor", 1))),
 			.crystals = zon.get(u32, "crystals", 0),
-			.erosion = zon.get(f32, "erosion", 0.5),
+			.soilCreep = zon.get(f32, "soilCreep", 0.5),
 			.minHeight = zon.get(i32, "minHeight", std.math.minInt(i32)),
 			.maxHeight = zon.get(i32, "maxHeight", std.math.maxInt(i32)),
 			.supportsRivers = zon.get(bool, "rivers", false),
@@ -464,9 +464,9 @@ pub const BlockStructure = struct { // MARK: BlockStructure
 		allocator.free(self.structure);
 	}
 
-	pub fn addSubTerranian(self: BlockStructure, chunk: *ServerChunk, startingDepth: i32, minDepth: i32, slope: i32, erosion: f32, x: i32, y: i32, seed: *u64) i32 {
+	pub fn addSubTerranian(self: BlockStructure, chunk: *ServerChunk, startingDepth: i32, minDepth: i32, slope: i32, soilCreep: f32, x: i32, y: i32, seed: *u64) i32 {
 		var depth = startingDepth;
-		var remainingSkippedBlocks = @as(i32, @intFromFloat(@as(f32, @floatFromInt(slope))*erosion)) - 1;
+		var remainingSkippedBlocks = @as(i32, @intFromFloat(@as(f32, @floatFromInt(slope))*soilCreep)) - 1;
 		for(self.structure) |blockStack| {
 			const total = blockStack.min + main.random.nextIntBounded(u32, seed, @as(u32, 1) + blockStack.max - blockStack.min);
 			for(0..total) |_| {