--- a/src/server/terrain/chunkgen/TerrainGenerator.zig
+++ b/src/server/terrain/chunkgen/TerrainGenerator.zig
@@ -81,18 +81,15 @@ pub fn generate(worldSeed: u64, chunk: *main.chunk.ServerChunk, caveMap: CaveMap
 						};
 
 						const surfaceBlock = caveMap.findTerrainChangeAbove(x, y, z) - chunk.super.pos.voxelSize;
-						var sum: i32 = 0;
-						var num: f32 = 0;
+						var slope: i32 = std.math.maxInt(i32);
 						for(cardinalDirections) |direction| {
 							const diff = if(caveMap.isSolid(x + direction[0], y + direction[1], z + direction[2]))
 								caveMap.findTerrainChangeAbove(x + direction[0], y + direction[1], z + direction[2]) - chunk.super.pos.voxelSize - surfaceBlock
 							else
 								caveMap.findTerrainChangeBelow(x + direction[0], y + direction[1], z + direction[2]) - surfaceBlock;
 
-							sum += @intCast(@abs(diff));
-							num += 1.0;
+							slope = @min(@as(i32, @intCast(@abs(diff))), slope);
 						}
-						const slope = @as(f32, @floatFromInt(sum))/num;
 						const erosion: f32 = biome.erosion;
 						var bseed: u64 = random.initSeed3D(worldSeed, .{chunk.super.pos.wx + x, chunk.super.pos.wy + y, chunk.super.pos.wz + z});
 						const airBlockBelow = caveMap.findTerrainChangeBelow(x, y, z);