Fixes #188