@careeoki 2025-07-21
Looks great, cliffs actually look like cliffs
<img width="1779" height="997" alt="image" src="https://github.com/user-attachments/assets/2a405ce5-4466-45ad-8b1e-e74931fb3a2c" />
<img width="1602" height="919" alt="image" src="https://github.com/user-attachments/assets/0787d48f-c165-4506-9593-e4b63701b262" />