@IntegratedQuantum 2025-07-23
That does make it rather ineffective for cliffs, since often while the cliff is steep, there will be a few blocks at similar height.