@codemob-dev 2025-07-24
> That does make it rather ineffective for cliffs, since often while the cliff is steep, there will be a few blocks at similar height.

This should be fixed now.