@codemob-dev 2025-07-24
> Also still waiting on that performance comparison.

With the PR: (
chunkgen: 37236 µs/task
chunkgen: 41582 µs/task

Without the PR: