@codemob-dev 2025-07-24
> Also still waiting on that performance comparison.

2 trials on each version, fully loading the spawn without moving.

With the PR: 
chunkgen: 37236 µs/task
chunkgen: 41582 µs/task

Without the PR:
chunkgen: 43120 µs/task
chunkgen: 51151 µs/task

The difference doesn't seem to be noticeable, I doubt that the PR improved performance so I'm guessing that there is too much variation to tell the difference.