@IntegratedQuantum 2025-07-24
These numbers seem suspiciously high, did you compile in release?