@codemob-dev 2025-07-24
I did, my computer just tends to be bad at chunkgen.