@codemob-dev 2025-07-25
> Performance impact is indeed quite small, I profiled it to be ~2 s out of ~84 s for the entire TerrainGenerator function.

Nice! The system can probably be expanded to help with #558, but thats probably for another PR.