@codemob-dev 2025-07-25
> Could https://github.com/PixelGuys/Cubyz/issues/588 be solved with a negative value maybe?

It would require some modifications of the code though, but yeah it probably could be.