@IntegratedQuantum 2025-07-25
Yeah, we don't want those